+++ src/farn/__init__.py
@@ -1,10 +1,12 @@
-from .farn import (  # noqa: F401
-    Case as Case,
-    Parameter as Parameter,
-    create_case_folders as create_case_folders,
-    create_case_list_files as create_case_list_files,
-    create_param_dict_files as create_param_dict_files,
-    create_samples as create_samples,
-    execute_command_set as execute_command_set,
+# from farn.core import Case as Case
+# from farn.core import Parameter as Parameter
+
+from .farn import (
     run_farn as run_farn,
+    create_samples as create_samples,
+    create_cases as create_cases,
+    create_case_folders as create_case_folders,
+    create_param_dict_files as create_param_dict_files,
+    create_case_list_files as create_case_list_files,
+    execute_command_set as execute_command_set,
 )